Extending resource schema (#10)
Extending resource schema and adding a counter that counts number of
consequential failed atempts to establish connection with the device.
Once defined threshold is reached, device is reported to be in `DOWN`
state.

This should help monitoring service to be a more robust solution for the
devices that are located on sites with bad connectivity.